Task 2 Part 2
Implemented file-based storage for transactions.
Added the saveTransaction(String transactionDetails) method inside the Transaction class, so every borrowing and returning transaction will be saved into a transactions.txt file. borrowBook and returnBook methods updated with a call to this method to persist the history of transactions in case of program exit.