Right Task 3 Part 3
Implemented `testSingletonTransaction` to test Singleton behavior of `Transaction` class using reflection. It checks the private constructor and that the instance using reflection and `Transaction.getTransaction()` returns the same object. Added assertions for constructor visibility and Singleton instance comparison. I also check all 3 tests at the same time if they are correct or not.